+++ HISTORY.md
@@ -13,7 +13,7 @@
     * new "Sarron" costume series, thanks, Meghan and Brian!
     * new Armenian translation, yay! Thanks to the contributors!
     * Expanded the `#open:` URL scheme to accept image URLs.
-    * new "ide_translation_dict" extension primitive
+    * new "ide_translation_dict" and "ide_set_translation_dict(data)" extension primitives
 * **Notable Changes:**
     * "this script" reporter has been morphed into the new general "this" introspection/self reflection reporter
     * RUN/CC and CALL/CC primitives have been deprecated and removed from the palette, will still be present and work in existing projects
@@ -44,6 +44,9 @@
     * German
     * Greek, thanks, HM100!
     * Armenian, thanks to the contributors!
+
+### 2023-05-08
+* extensions: added "ide_set_translation_dict(data)" extension primitive
 
 ### 2023-05-08
 * gui: reformatted some code to please JSHint ;)